added snake and snake_decoder
snake decoder is just the 64 bit decoder
snake.v is how the snake length updates ig? 
i see the ball boundary thing and i think i have the idea for it. it covers the entire board as an array (explodes)
that PROBABLY means i have to change my snake file